\name{quic}
\alias{quic}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
quadratic function outputs values that increase in a quadratic manner from 1 to the number put into the function
}
\description{
quadratic function outputs values that increase in a quadratic manner from 1 to the number put into the function
}
\usage{
quic(Item_count)
}


\details{
Ideal to model an Item Positions Effect for fixed links modelling in lavaan
}
\value{
list of values
}
\references{
Troche, S. J., Wagner, F. L., Schweizer, K., & Rammsayer, T. H. (2016). The Structural Validity of the Culture Fair Test Under Consideration of the Item-Position Effect. European Journal of Psychological Assessment, 35(2), 182–189. https://doi.org/10.1027/1015-5759/a000384
}
\author{
Helene M. von Gugelberg
}
\note{
%%  ~~further notes~~
}



\examples{

quic(18)

}

